+++ play-frontend-hmrc-play-30/src/test/scala/uk/gov/hmrc/hmrcfrontend/views/components/HmrcHeaderSpec.scala
@@ -50,7 +50,7 @@ class HmrcHeaderSpec extends TemplateUnitSpec[Header, HmrcHeader]("hmrcHeader") 
       val componentTry = Try(hmrcHeader(Header()))
 
       componentTry.isSuccess shouldBe true
-      componentTry.get.body    should include("M22.6 10.4c-1")
+      componentTry.get.body    should include("M33.1,9.8c.2")
     }
 
     """display St Edwards crown logo when set by config""" in {
@@ -64,7 +64,7 @@ class HmrcHeaderSpec extends TemplateUnitSpec[Header, HmrcHeader]("hmrcHeader") 
       val componentTry = Try(hmrcHeader(Header()))
 
       componentTry.isSuccess shouldBe true
-      componentTry.get.body    should include("M6.7 12.2c1")
+      componentTry.get.body    should include("M13.4,22.3c2")
     }
 
     """display Tudor crown when no config is found""" in {
@@ -74,7 +74,21 @@ class HmrcHeaderSpec extends TemplateUnitSpec[Header, HmrcHeader]("hmrcHeader") 
       val componentTry = Try(hmrcHeader(Header()))
 
       componentTry.isSuccess shouldBe true
-      componentTry.get.body    should include("M22.6 10.4c-1")
+      componentTry.get.body    should include("M33.1,9.8c.2")
+    }
+  }
+
+  "output of rebrand enabled" should {
+    "match output of rebrand disabled" when {
+      "rebrand is enabled by argument" in {
+        val appWithRebrand    = buildAnotherApp(Map("play-frontend-hmrc.useRebrand" -> "true"))
+        val appWithoutRebrand = buildAnotherApp(Map("play-frontend-hmrc.useRebrand" -> "false"))
+
+        val viewWithRebrand    = appWithRebrand.injector.instanceOf[HmrcHeader]
+        val viewWithoutRebrand = appWithoutRebrand.injector.instanceOf[HmrcHeader]
+
+        viewWithRebrand.apply(Header()) shouldBe viewWithoutRebrand.apply(Header(rebrand = Some(true)))
+      }
     }
   }
 }