+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 For compatibility information see `govukFrontendVersion` and `hmrcFrontendVersion` in
 [LibDependencies](project/LibDependencies.scala)
 
+## [12.2.0] - 2025-06-03
+
+### Changed
+
+- Updated `hmrc-frontend` to v6.68.0
+- Updated `govuk-frontend` to v5.10.2
+- This introduces GOV.UK rebrand behind feature flag
+
 ## [12.1.0] - 2025-05-08
 
 ### Changed